Fix more signed/unsigned integer mismatch warnings, fix few problems to build on windows
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@4179 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/tools/imagetools/ImageObject.C
@@ -133,7 +133,7 @@ ImageObject::Write(const char *filename)
 
     // Check the extension.
     bool retval = true;
-    int len = strlen(filename);
+    int len = (int) strlen(filename);
     int writeType = 0;
     if(len > 4)
     {
